Many mailphoto improvements, mailmaint newline
This adds a whole bunch of comments and the ability to specify the
photo instead of always taking it from a webcam.

This is untested, but I was hoping to only commit the mailmaint
changes (just a newline at the end) so I guess I'm stuck now.  Maybe I
should learn to use git at some point :).