Get all randompass data from /dev/random
I'm a bit worried that the other way was actually insecure so instead
I'm just going to fetch a big block of data from /dev/random and print
it out to the screen.  I probably should have done this a long time
ago, but now I'm resetting all my passwords so I think it's a good
time to.